requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.0.20
pydantic>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
# Action Executor — MCP integration
mcp>=1.0.0
langchain-mcp-adapters>=0.1.0
# API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
psycopg[binary]>=3.1.0
# Auth0 JWT
PyJWT>=2.8.0
httpx>=0.27.0
# Slack OAuth + Web API
slack_sdk>=3.27.0
# Celery + Redis task queue
celery[redis]>=5.3.0
redis>=5.0.0
# LangGraph Postgres checkpointer
langgraph-checkpoint-postgres>=2.0.0
